templates/python/requirements.txt
nose
mock
PyHamcrest
coverage
flake8
